+++ api/src/main/java/org/openmrs/module/eptsreports/reporting/calculation/pvls/PregnantCalculation.java
@@ -18,7 +18,7 @@ import org.openmrs.calculation.result.ListResult;
 import org.openmrs.module.eptsreports.metadata.HivMetadata;
 import org.openmrs.module.eptsreports.reporting.calculation.AbstractPatientCalculation;
 import org.openmrs.module.eptsreports.reporting.calculation.BooleanResult;
-import org.openmrs.module.eptsreports.reporting.calculation.EptsCalculations;
+import org.openmrs.module.eptsreports.reporting.calculation.common.EPTSCalculationServie;
 import org.openmrs.module.eptsreports.reporting.utils.EptsCalculationUtils;
 import org.openmrs.module.reporting.common.TimeQualifier;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -60,7 +60,7 @@ public class PregnantCalculation extends AbstractPatientCalculation {
     Set<Integer> femaleCohort = EptsCalculationUtils.female(cohort, context);
 
     CalculationResultMap pregnantMap =
-        EptsCalculations.getObs(
+        EPTSCalculationServie.getObs(
             pregnant,
             femaleCohort,
             Arrays.asList(location),
@@ -70,7 +70,7 @@ public class PregnantCalculation extends AbstractPatientCalculation {
             context);
 
     CalculationResultMap markedPregnantByWeeks =
-        EptsCalculations.getObs(
+        EPTSCalculationServie.getObs(
             pregnantBasedOnWeeks,
             femaleCohort,
             Arrays.asList(location),
@@ -80,7 +80,7 @@ public class PregnantCalculation extends AbstractPatientCalculation {
             context);
 
     CalculationResultMap markedPregnantDueDate =
-        EptsCalculations.getObs(
+        EPTSCalculationServie.getObs(
             pregnancyDueDate,
             femaleCohort,
             Arrays.asList(location),
@@ -90,10 +90,10 @@ public class PregnantCalculation extends AbstractPatientCalculation {
             context);
 
     CalculationResultMap markedPregnantInProgram =
-        EptsCalculations.allProgramEnrollment(ptv, femaleCohort, context);
+        EPTSCalculationServie.allProgramEnrollment(ptv, femaleCohort, context);
 
     CalculationResultMap lastVl =
-        EptsCalculations.lastObs(
+        EPTSCalculationServie.lastObs(
             Arrays.asList(labEncounterType, adultFollowup, pediatriaFollowup),
             viralLoadConcept,
             location,
